airflow/requirements.txt
apache-airflow-providers-google==7.0.0
pyarrow==8.0.0
pandas==1.4.2
yfinance==0.1.70